// angular-biometric-designer/src/app/shared/models/svcard/cardrenew.dto.ts
export class CardReNewDto {
    id: string = '';
    batch_num: string = '';
    product_code: string = '';
    product_name: string = '';
    cust_no: string = '';
    cust_type: string = '';
    cust_name: string = '';
    sub_cust_name: string = '';
    cust_phone: string = '';
    cust_mobile: string = '';
    unique_id_value: string = '';
    birth_day: string = '';
    address_1: string = '';
    address_2: string = '';
    address_3: string = '';
    address_4: string = '';
    country_code: string = '';
    city_code: string = '';
    user_branch: string = '';
    branch_code: string = '';
    agent_code: string = '';
    card_id: string = '';
    card_no: string = '';
    card_new_no: string = '';
    card_old_no: string = '';
    card_type: string = '';
    card_name: string = '';
    card_name_sub: string = '';
    release_form: string = '';
    service_code: string = '';
    service_name: string = '';
    reference_no: string = '';
    cust_ac_no: string = '';
    auth_id: string = '';
    auth_stat: string = '';
    auth_stat_text: string = '';
    created_by: string = '';
    created_date: string = '';
    approved_by: string = '';
    approved_date: string = '';
    unique_id_date: string = '';
    expired_unique_id_date: string = '';
    card_class: string = '';
    form_of_guarantee: string = '';
    primary_card_number: string = '';
    cardholder_rel: string = '';
    sub_unique_id_value: string = '';
    sub_date_gttt: string = '';
    sub_birth_day: string = '';
    credit_card_limit: string = '';
    registration_limit: string = '';
    debt_ratio: string = '';
    payment_account: string = '';
    register_online_payment: boolean = false;
    register_auto_debit: boolean = false;
    reason: string = '';
    credit_product_code: string = '';
}